requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.5
pika==1.3.2
